fix: provider 정리 및 검색 결과 중복 클릭 방지
