+++ src/types/dropdown.ts
@@ -1,10 +1,11 @@
 export interface DropdownType {
   dropdownRef: React.RefObject<HTMLUListElement>;
+  setTarget: React.Dispatch<React.SetStateAction<HTMLElement | null | undefined>>;
   searchListData: string[];
   inputText: string;
   isTotal: boolean;
   isSearchLoading: boolean;
-  handleSearchFetch: (type: string, inputText: string) => Promise<void>;
+  handleSubmit: (event: React.FormEvent, inputText: string) => Promise<void>;
 }
 
 export interface HighlightType {
@@ -13,8 +14,12 @@ export interface HighlightType {
 }
 
 export interface InputTodoType {
+  inputText: string;
+  setInputText: React.Dispatch<React.SetStateAction<string>>;
   inputRef: React.RefObject<HTMLInputElement>;
+  isAddLoading: boolean;
   setInputFocus: () => void;
   handleInputClick: () => void;
   handleSearchFetch: (type: string, inputText: string) => Promise<void>;
+  handleSubmit: (event: React.FormEvent, inputText: string) => Promise<void>;
 }
